requirements.txt
ultralytics==8.3.0
wandb==0.17.0
fastapi==0.111.0
uvicorn==0.30.0
onnxruntime==1.18.0
opencv-python==4.10.0.84
PyYAML==6.0.1
